Arrows.jl

Arrows.jl

Arrows.jl is a differentiable programming environment implemented in Julia. The goal is to combine the benefits of deep neural networks - namely, that they are differentiable, - with the benefits of modern programming languages - recursion, modularity, higher-orderness, types. To do this, we build upon the formalism of Arrows.

Installation

Arrows is built in Julia but not yet in the official Julia Package repository. You can still easily install it from a Julia repl with:

Pkg.clone("https://github.com/zenna/Arrows.jl.git")

Arrows.jl

Installation

Arrows.CompArrow — Type.

A Composite Arrow: An Arrow composed of multiple Arrows

source

Arrows.CompArrow